requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4

cffi~=1.15.1
pip~=23.1.2
attrs~=22.2.0
DateTime~=5.0
wheel~=0.38.4
cryptography~=41.0.2
keyring~=23.13.1
colorama~=0.4.6
requests~=2.28.1
setuptools~=58.1.0
uritemplate~=4.1.1
google-api-python-client~=2.80.0
schedule~=1.2.0
django-scheduler~=0.10.1
pytils~=0.4.1